moves file to helper directory
